wicked_zerg_challenger/requirements.txt
burnysc2>=5.0.0
numpy>=1.20.0
matplotlib>=3.3.0
psutil>=5.8.0
s2clientprotocol>=4.19.0.0
